CORE-662 extract hard-coded data for compiler
This change extracts hard-coded array methods and keywords from `server/src/keywordCompletion.ts` and places them in `data/arrayMethods.json`, and `data/keywordToItemKind.json`, respectively.

We hard-code the type of the import that specifies part of the `CompletionItemKind` with

```
{
  [keyword: string] : "Text" | "Method" | "Function" |
    "Constructor" | "Field" | "Variable" | "Class" |
    "Interface" | "Module" | "Property" | "Unit" | "Value" |
    "Enum" | "Keyword" | "Snippet" | "Color" | "File" |
    "Reference" | "Folder" | "EnumMember" | "Constant" |
    "Struct" | "Event" | "Operator" | "TypeParameter"
}
```

because that avoids an error from TypeScript, later, that complains about invalid index signatures.