give `App` its own doc and item kind for autocompl
Without this change, when typing "." after typing "Reach", "App" shows up as a suggestion, but with no documentation, and its CompletionKind is just text.

We add this change so that when typing "." after typing "Reach", "App" in the CompletionItem suggestions dropdown list has documentation, and appears as a function, as it should.