CORE-887 add new info to suggestions dropdown list
This change has the effect of changing the righthand side of the suggestions dropdown list when developers select a Reach keyword.

For example, the righthand side of `checkCommitment` now says `(Function)`, instead of `checkCommitment`. So, the suggestions dropdown list now adds new information, instead of just repeating existing information, for Reach keywords.