change q-and-a background color
Currently `q-and-a` blocks look identical to `note` blocks.  This is
particularly an issue for “true or false” statements that are not
clearly questions.  The tutorial has `q-and-a` blocks that make false
statements but that look identical to `note` blocks which offer
authoritative guidance.  I'm not going to suggest that I have the best
aesthetic sense for what the style should be, but it should absolutely
be different from `note`.

Ideally `q-and-a` blocks could also come with an explicit label that
shows that they have a question, not just implicit styling to show
that it's a question.