Fix seclinks with empty titles.
We have some section headings where (part of) the title of the section
is in backticks, like `this`.  When processing section headings like
that that define anchors for `seclink`, the backtick splits the text
into different list elements.  Previously we only used the first
element in a list to generate the title used for references, and thus
for `seclink`.  This resulted in links with empty text in the anchor.
The links could be discovered using the web inspector, but were
invisible and unclickable on the page.  Also, everywhere `seclink` is
used, some text output is expected gramatically at the use site.

This patch changes the processing to include backticked words and text
following them to be included in the title.  This makes previously
empty `seclink`s visible, legible, and clickable.

It also adds a couple of guard rails with `fail` if any empty titles
for xrefs make it through this improved processing in the future.