docs: add externalRef form
This allows, for example, Javascript code blocks to consistently link
to external reference material, such as the MDN or React
documentation.

This commit includes a few externalRef uses mostly to test it.  We
could add an exhaustive number of externalRefs, but that's probably a
lot of work.  I'm not sure which identifiers would be most helpful to
link.

While `externalRef` augments the reference database, it returns
nothing to the document itself, and thus has no immediate effect on
the containing document.  So it doesn't really matter which page these
are in to make the effect.  I stuffed these references in the search
page, because it seemed like the most reasonable place to stuff them
in without making a blank new page.