Add solidity reserved words
I'm not sure these are technically reserved words in the Solidity language, but they are keys in the resulting contract objects.
Without adding these to the list, they can be used as view names.
When used as view names, they conflict with other keys in the contract object and break.