Add initial Wisdom For Sale .rsh and .mjs files
This ran properly in Ubuntu and output the role of seller and the network as Algo as intended.

This is intended to be a test of the process of creating a branch off master in GitHub Desktop separated from my forked branch, but adds files required for Wisdom For Sale.