get raw api-many test working for algorand with 20 args
IE - the signature string printed in the backend for algorand now
turns long signatures into signatures with a tuple at argument 15.

That said, I'm not sure how I feel about the solution of packing the
signature.  To make a raw call you effectively need to know where the
argument list split is to manually tuple the later arguments.
So... it seems like from the perspective of making raw calls that
functions of 16+ arguments just don't exist and you need to know about
the need to pack in tuples.

The standard doesn't really make it explicit whether the public
signature (eg. the one you will see when loading it with algosdk) for
a 16+ arg function should be the original signature or converted
signature.  My feeling is that it SHOULD be the original, and the
packing is just an implementation detail.  It would be helpful and
telling if algosdk had a function to encode a whole argument array at
once, handling this tupling automatically (or not) and whichever way
it works could guide our design.  But I don't see such a function.
The way I'm manually mapping the encoding function from the formal
parameter definition to the arguments definitely favors the
publication of the tuple version signature or then you have to
manually split both the arguments and the formal parameters.  But
making everything split and make a tuple makes it feel to me like
algorand does NOT support 16+ argument functions and merely has a
workaround, while the standard language makes me feel like signatures
can have more than 16 arguments.  The algosdk doesn't throw an
exception if you load a method from a signature with more than 16
arguments.  So... that's a point in favor of leaving the original
signature.

So... I feel like I would rather keep the original signature.

But if we do keep this behavior, now there's yet another place in
ALGO.hs that splits arguments at a particular place.  Are they perhaps
constructed and/or flow into ALGO.hs in one particular spot so we can
do this split just once?