Fix flaky api-multiple-calls test... maybe.
I have yet to be able to reproduce the failure locally.  But my best
guess is that there was a race condition where between the
`A.interact.deployed()` call and the publish/commit pair after it the
client had a window of opportunity to perform an API call before it
was in the correct state.  This removes that extra publish/commit
pair, so the deploy interaction really does mean that it's in the
correct state.