fix verification for multiplication overflow where arg might be zero
This fixes the bug with verification of multiplication overflow.  The problem is that verification that overflow doesn't happen relies on using division, but SMT defines division by zero as zero instead of an error.  This allows a situation where if the second argument of a multiplication can be zero then it triggers division by zero, leading to failure to prove the desired fact about the division.  This adds a check for whether the second argument is zero and does a logical OR of the zero test and the division test.